+++ packages/stream_video_flutter/lib/stream_video_flutter.dart
@@ -34,6 +34,7 @@ export 'src/call_participants/local_video.dart';
 export 'src/call_screen/call_container.dart';
 export 'src/call_screen/call_content/call_app_bar.dart';
 export 'src/call_screen/call_content/call_content.dart';
+export 'src/call_screen/call_content/picture_in_picture/android_pip_manager.dart';
 export 'src/call_screen/call_content/picture_in_picture/android_pip_overlay.dart';
 export 'src/call_screen/call_content/picture_in_picture/picture_in_picture_configuration.dart';
 export 'src/call_screen/call_content/picture_in_picture/stream_picture_in_picture_android_view.dart';